+++ android/src/main/java/com/twiliorn/library/RNVideoViewGroup.java
@@ -10,7 +10,7 @@ import android.content.Context;
 import android.graphics.Point;
 import android.view.View;
 import android.view.ViewGroup;
-import androidx.annotation.StringDef;
+import android.support.annotation.StringDef;
 
 import com.facebook.react.bridge.WritableMap;
 import com.facebook.react.bridge.WritableNativeMap;
