+++ android/src/main/java/com/twiliorn/library/TwilioRemotePreviewManager.java
@@ -7,7 +7,7 @@
 
 package com.twiliorn.library;
 
-import androidx.annotation.Nullable;
+import android.support.annotation.Nullable;
 import android.util.Log;
 
 import com.facebook.react.common.MapBuilder;
@@ -52,6 +52,7 @@ public class TwilioRemotePreviewManager extends SimpleViewManager<TwilioRemotePr
 
     @ReactProp(name = "trackSid")
     public void setTrackId(TwilioRemotePreview view, @Nullable String trackSid) {
+
         Log.i("CustomTwilioVideoView", "Initialize Twilio REMOTE");
         Log.i("CustomTwilioVideoView", trackSid);
         myTrackSid = trackSid;
@@ -60,7 +61,7 @@ public class TwilioRemotePreviewManager extends SimpleViewManager<TwilioRemotePr
 
     @ReactProp(name = "applyZOrder", defaultBoolean = false)
     public void setApplyZOrder(TwilioRemotePreview view, boolean applyZOrder) {
-      view.applyZOrder(applyZOrder);
+        view.applyZOrder(applyZOrder);
     }
 
     @Override
@@ -70,12 +71,12 @@ public class TwilioRemotePreviewManager extends SimpleViewManager<TwilioRemotePr
 
     @Override
     public Map getExportedCustomBubblingEventTypeConstants() {
-      return MapBuilder.builder()
-          .put(
-            ON_FRAME_DIMENSIONS_CHANGED,
-              MapBuilder.of(
-                  "phasedRegistrationNames",
-                  MapBuilder.of("bubbled", ON_FRAME_DIMENSIONS_CHANGED)))
-                  .build();
-  }
+        return MapBuilder.builder()
+                .put(
+                        ON_FRAME_DIMENSIONS_CHANGED,
+                        MapBuilder.of(
+                                "phasedRegistrationNames",
+                                MapBuilder.of("bubbled", ON_FRAME_DIMENSIONS_CHANGED)))
+                .build();
+    }
 }
