+++ android/src/main/java/com/twiliorn/library/TwilioVideoPreviewManager.java
@@ -7,7 +7,7 @@
 
 package com.twiliorn.library;
 
-import androidx.annotation.Nullable;
+import android.support.annotation.Nullable;
 
 import com.facebook.react.bridge.ReadableArray;
 import com.facebook.react.common.MapBuilder;
@@ -32,11 +32,11 @@ public class TwilioVideoPreviewManager extends SimpleViewManager<TwilioVideoPrev
 
     @ReactProp(name = "scaleType")
     public void setScaleType(TwilioVideoPreview view, @Nullable String scaleType) {
-      if (scaleType.equals("fit")) {
-        view.setScalingType(RendererCommon.ScalingType.SCALE_ASPECT_FIT);
-      } else {
-        view.setScalingType(RendererCommon.ScalingType.SCALE_ASPECT_FILL);
-      }
+        if (scaleType.equals("fit")) {
+            view.setScalingType(RendererCommon.ScalingType.SCALE_ASPECT_FIT);
+        } else {
+            view.setScalingType(RendererCommon.ScalingType.SCALE_ASPECT_FILL);
+        }
     }
 
     @ReactProp(name = "scalesType")
@@ -51,7 +51,7 @@ public class TwilioVideoPreviewManager extends SimpleViewManager<TwilioVideoPrev
 
     @ReactProp(name = "applyZOrder", defaultBoolean = true)
     public void setApplyZOrder(TwilioVideoPreview view, boolean applyZOrder) {
-      view.applyZOrder(applyZOrder);
+        view.applyZOrder(applyZOrder);
     }
 
     @Override
